renamed package from generic view name (is also confusing because mvc) to web
