+++ src/main/java/edu/ipd/kit/crowdcontrol/proto/controller/TaskController.java
@@ -11,9 +11,9 @@ import edu.ipd.kit.crowdcontrol.proto.databasemodel.tables.pojos.Experiment;
 import edu.ipd.kit.crowdcontrol.proto.databasemodel.tables.pojos.Hit;
 import edu.ipd.kit.crowdcontrol.proto.databasemodel.tables.records.RatingsRecord;
 import edu.ipd.kit.crowdcontrol.proto.json.JSONRating;
-import edu.ipd.kit.crowdcontrol.proto.view.CreativeTaskView;
-import edu.ipd.kit.crowdcontrol.proto.view.RatingTaskView;
-import edu.ipd.kit.crowdcontrol.proto.view.TaskView;
+import edu.ipd.kit.crowdcontrol.proto.web.CreativeTaskView;
+import edu.ipd.kit.crowdcontrol.proto.web.RatingTaskView;
+import edu.ipd.kit.crowdcontrol.proto.web.TaskView;
 import org.jooq.DSLContext;
 import org.jooq.Record1;
 import org.jooq.impl.DSL;
