+++ src/main/java/edu/ipd/kit/crowdcontrol/proto/web/FreeMarkerEngine.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package edu.ipd.kit.crowdcontrol.proto.view;
+package edu.ipd.kit.crowdcontrol.proto.web;
 
 import freemarker.template.Configuration;
 import freemarker.template.Template;
